requirements.txt
Flask == 1.0.2
google-cloud-datastore==1.7.1
Werkzeug == 0.15.2
gunicorn == 19.9.0
python-dotenv == 0.10.3
google-api-python-client == 1.7.9